Fix #3793 part two, panels hidden on variant view
