+++ scout/server/blueprints/variant/utils.py
@@ -148,6 +148,24 @@ def update_transcripts_information(variant_gene, hgnc_gene, variant_obj, genome_
         transcript["change_str"] = transcript_str(transcript, hgnc_symbol)
 
 
+def update_variant_case_panels(store, case_obj, variant_obj):
+    """Populate variant with case gene panels with info on e.g. if a panel was removed on variant_obj.
+    Variant objects panels are only a list of matching panel names.
+
+    Args:
+        store(adapter.MongoAdapter)
+        case_obj(dict)
+        variant_obj(dict)
+    """
+
+    variant_panel_names = variant_obj.get("panels", [])
+    case_panel_objs = [
+        panel for panel in case_obj.get("panels", []) if panel["panel_name"] in variant_panel_names
+    ]
+
+    variant_obj["case_panels"] = case_panel_objs
+
+
 def add_gene_info(store, variant_obj, gene_panels=None, genome_build=None):
     """Adds information to variant genes from hgnc genes and selected gene panels.
